README.md added

+++ Badges_Console/ProgramUI.cs
@@ -9,7 +9,11 @@ namespace Badges_Console
 {
     class ProgramUI
     {
-        private readonly BadgeRepository _badgeRepo = new BadgeRepository();
+        private readonly IBadgeRepository _badgeRepo;
+        public ProgramUI(IBadgeRepository repo)
+        {
+            _badgeRepo = repo;
+        }
 
         public void Run()
         {
@@ -65,6 +69,10 @@ namespace Badges_Console
 
             Console.WriteLine("List all doors this badge needs access to(COMMA SEPARATED): ");
             newbadge.Doors = Console.ReadLine().Split(',').ToList();
+            for (int i = 0; i < newbadge.Doors.Count; i++)
+            {
+                newbadge.Doors[i] = newbadge.Doors[i].Trim();
+            }
 
             if (_badgeRepo.CreateBadge(newbadge))
                 Console.WriteLine("New badge created");
@@ -73,37 +81,47 @@ namespace Badges_Console
         }
         private void UpdateDoorsInBadge()
         {
-            Console.WriteLine("UPDATING DOORS IN A BADGE\n");
-            Console.WriteLine("What would you like to do?\n" +
-                              "1. Remove a door\n" +
-                              "2. Add a door\n");
-
-            char choice = Console.ReadKey().KeyChar;
-            Console.WriteLine();
-            switch(choice)
+            if(_badgeRepo.GetBadges.Count > 0)
             {
-                case '1':
-                    Console.Write("Enter badge id to remove a door: ");
-                    int idToRemove = int.Parse(Console.ReadLine());
-                    Console.Write("Enter a door to remove access: ");
-                    if (_badgeRepo.RemoveDoor(idToRemove, Console.ReadLine()))
-                        Console.WriteLine("Door removed!");
-                    else
-                        Console.WriteLine("Door is not removed");
-                    break;
-                case '2':
-                    Console.Write("Enter badge id to add a door: ");
-                    int idToAdd = int.Parse(Console.ReadLine());
-                    Console.Write("Enter a door to add access: ");
-                    if (_badgeRepo.AddDoor(idToAdd, Console.ReadLine()))
-                        Console.WriteLine("Door added!");
-                    else
-                        Console.WriteLine("Door is not added");
-                    break;
-                default:
-                    Console.Clear();
-                    UpdateDoorsInBadge();
-                    break;
+                Console.WriteLine("UPDATING DOORS IN A BADGE\n");
+                ShowAllBadges();
+                Console.WriteLine();
+                Console.WriteLine("What would you like to do?\n" +
+                                  "1. Remove a door\n" +
+                                  "2. Add a door\n");
+
+                char choice = Console.ReadKey().KeyChar;
+                Console.WriteLine();
+
+                switch (choice)
+                {
+                    case '1':
+                        Console.Write("Enter badge id to remove a door: ");
+                        int idToRemove = int.Parse(Console.ReadLine());
+                        Console.Write("Enter a door to remove access: ");
+                        if (_badgeRepo.RemoveDoor(idToRemove, Console.ReadLine()))
+                            Console.WriteLine("Door removed!");
+                        else
+                            Console.WriteLine("Door is not removed");
+                        break;
+                    case '2':
+                        Console.Write("Enter badge id to add a door: ");
+                        int idToAdd = int.Parse(Console.ReadLine());
+                        Console.Write("Enter a door to add access: ");
+                        if (_badgeRepo.AddDoor(idToAdd, Console.ReadLine()))
+                            Console.WriteLine("Door added!");
+                        else
+                            Console.WriteLine("Door is not added");
+                        break;
+                    default:
+                        Console.Clear();
+                        UpdateDoorsInBadge();
+                        break;
+                }
+            }
+            else
+            {
+                Console.WriteLine("There are no badges to update");
             }
         }
         private void DeleteDoorsFromBadge()
@@ -111,6 +129,8 @@ namespace Badges_Console
             if (_badgeRepo.GetBadges.Count > 0)
             {
                 Console.WriteLine("DELETING ALL DOORS FROM A BADGE\n");
+                ShowAllBadges();
+                Console.WriteLine();
                 Console.Write("Enter Badge ID:");
                 int id = int.Parse(Console.ReadLine());
 
@@ -135,7 +155,7 @@ namespace Badges_Console
                     Console.CursorLeft = col_DoorAccess;
                     foreach (var door in badge.Doors)
                     {
-                        Console.Write(door.Trim() + " ");
+                        Console.Write(door + " ");
                     }
                     Console.WriteLine();
                 }
